refactor: PHPStan al nivel 10 (maximo) — niveles 1-8 limpios sin excepciones
NIVELES 7 Y 8 GENUINOS (~30 problemas reales corregidos):
- strtotime() puede devolver false y se pasaba a date(): casteo explicito
  en finanzas, gastos, tablero, portal y helpers de fecha
- max() sobre array_column() potencialmente vacio en tablero y finanzas
- filas SQL usadas sin comprobar existencia: guardas en configuracion
  (usuario borrado -> cerrarSesion) y receta (recarga tras update)
- preg_replace() con $_POST['telefono'] que un formulario manipulado
  puede enviar como array
- session_name() y json_encode() pueden devolver false (setcookie/curl)
- getNombreProducto() y generarNumeroLote() castean fetchColumn() a string

- redirigir() y cerrarSesion() se declaran ": never". Ademas de ser
  correcto (ambas terminan en exit), permite a PHPStan entender el flujo
  y elimino por si solo varios falsos positivos de acceso a datos.

NIVELES 9-10 CON BASELINE:
- level: 10 + phpstan-baseline.neon con las 886 ocurrencias heredadas
  (casi todas filas SQL array<string, mixed> pasadas a funciones tipadas)
- Verificado que NO es un "ignorar todo": al introducir una violacion
  nueva de nivel 10, el analisis falla; al quitarla, vuelve a verde.
  Es decir, todo codigo nuevo debe ser nivel-10 limpio o el CI falla.

FALSO POSITIVO DOCUMENTADO:
- PDO::prepare()/query() se declaran "PDOStatement|false" en PHP, pero
  config/db.php usa ERRMODE_EXCEPTION: la rama false es inalcanzable.
  Se probo antes con stubFiles y no funciona (PHPStan da prioridad a sus
  stubs del core), asi que se excluye por identificador+mensaje exacto
  con la justificacion escrita en la configuracion.

Verificado: analisis en verde, suite 122 pruebas / 248 aserciones.
Version 1.6.0 sincronizada en CHANGELOG, APP_VERSION y README.

+++ controllers/GastoController.php
@@ -126,7 +126,7 @@ class GastoController {
         for ($i = 6; $i >= 0; $i--) {
             $f = date('Y-m-d', strtotime("-$i days"));
             $gastos_7d[] = [
-                'lbl' => date('d/m', strtotime($f)),
+                'lbl' => date('d/m', (int) strtotime($f)),
                 'v'   => (float)($gastos_7d_raw[$f] ?? 0)
             ];
         }
